Add `TopStoriesViewController` snapshot spec
Made some important changes on how `TopStoryViewModel` is making use of
RxSwift. I should avoid as much as possible go into the imperative
world, better to go along with the framework.

The snapshot is working but I'm not happy with the setup. It is too
long and complex, and I don't even know if I'm doing it properly.